+++ tests/test_predicates.py
@@ -25,26 +25,29 @@ import unittest
 from repoze.what import predicates
 
 
+class BasePredicateTester(unittest.TestCase):
+    """Base test case for predicates."""
+    
+    def eval_met_predicate(self, p, environ):
+        """Evaluate a predicate that should be met"""
+        self.assertEqual(p.eval_with_environ(environ), None)
+    
+    def eval_unmet_predicate(self, p, environ, expected_error):
+        """Evaluate a predicate that should not be met"""
+        try:
+            p.eval_with_environ(environ)
+        except predicates.PredicateError, error:
+            self.assertEqual(unicode(error), expected_error)
+
+
 #{ The test suite itself
 
 
-class TestPredicate(unittest.TestCase):
+class TestPredicate(BasePredicateTester):
     
     def test_eval_with_environ_isnt_implemented(self):
         p = MockPredicate()
         self.failUnlessRaises(NotImplementedError, p.eval_with_environ, None)
-    
-    def test_errors_are_added_if_required(self):
-        p = EqualsTwo()
-        environ = {'test_number': 3}
-        p.eval_with_environ(environ)
-        self.assertEqual(p.error, p.message % p.__dict__)
-    
-    def test_errors_arent_added_if_not_required(self):
-        p = EqualsTwo()
-        environ = {'test_number': 2}
-        p.eval_with_environ(environ)
-        self.assertEqual(p.error, '')
     
     def test_message_is_changeable(self):
         previous_msg = EqualsTwo.message
@@ -61,11 +64,10 @@ class TestPredicate(unittest.TestCase):
         unicode_msg = u'请登陆'
         p = EqualsTwo(msg=unicode_msg)
         environ = {'test_number': 3}
-        p.eval_with_environ(environ)
-        self.assertEqual(p.error, unicode_msg)
+        self.eval_unmet_predicate(p, environ, unicode_msg)
 
 
-class TestCompoundPredicate(unittest.TestCase):
+class TestCompoundPredicate(BasePredicateTester):
     
     def test_one_predicate_works(self):
         p = EqualsTwo()
@@ -79,93 +81,81 @@ class TestCompoundPredicate(unittest.TestCase):
         self.assertEqual(cp.predicates, (p1, p2))
 
 
-class TestNotPredicate(unittest.TestCase):
+class TestNotPredicate(BasePredicateTester):
     
     def test_success(self):
         environ = {'test_number': 4}
-        # It must not be greater than 5
+        # It must NOT equal 4
         p = predicates.Not(EqualsFour())
-        # It's greater than 5!
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error, 'The condition must not be met')
+        # It equals 4!
+        self.eval_unmet_predicate(p, environ, 'The condition must not be met')
+    
+    def test_success_with_custom_message(self):
+        environ = {'test_number': 4}
+        # It must not equal 4
+        p = predicates.Not(EqualsFour(), msg='It must not equal four')
+        # It equals 4!
+        self.eval_unmet_predicate(p, environ, 'It must not equal four')
     
     def test_failure(self):
-        environ = {'test_number': 7}
-        # It must not be less than 5
-        p = predicates.Not(LessThan(5))
-        # It's greater than 5!
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
-    
-    def test_custom_message(self):
         environ = {'test_number': 4}
-        # It must not be greater than 5
-        p = predicates.Not(EqualsFour(), msg='It must not equal four')
-        # It's greater than 5!
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error, 'It must not equal four')
+        # It must NOT be greater than 5
+        p = predicates.Not(LessThan(5))
+        # It's less than 5!
+        self.eval_met_predicate(p, environ)
 
 
-class TestAllPredicate(unittest.TestCase):
+class TestAllPredicate(BasePredicateTester):
     
     def test_one_true(self):
         environ = {'test_number': 2}
         p = predicates.All(EqualsTwo())
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
         
     def test_one_false(self):
         environ = {'test_number': 3}
         p = predicates.All(EqualsTwo())
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error, "Number 3 doesn't equal 2")
+        self.eval_unmet_predicate(p, environ, "Number 3 doesn't equal 2")
     
     def test_two_true(self):
         environ = {'test_number': 4}
         p = predicates.All(EqualsFour(), GreaterThan(3))
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
     
     def test_two_false(self):
         environ = {'test_number': 1}
         p = predicates.All(EqualsFour(), GreaterThan(3))
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error, "Number 1 doesn't equal 4")
+        self.eval_unmet_predicate(p, environ, "Number 1 doesn't equal 4")
     
     def test_two_mixed(self):
         environ = {'test_number': 5}
         p = predicates.All(EqualsFour(), GreaterThan(3))
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error, "Number 5 doesn't equal 4")
+        self.eval_unmet_predicate(p, environ, "Number 5 doesn't equal 4")
 
 
-class TestAnyPredicate(unittest.TestCase):
+class TestAnyPredicate(BasePredicateTester):
     
     def test_one_true(self):
         environ = {'test_number': 2}
         p = predicates.Any(EqualsTwo())
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
         
     def test_one_false(self):
         environ = {'test_number': 3}
         p = predicates.Any(EqualsTwo())
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error,
+        self.eval_unmet_predicate(p, environ, 
                          "At least one of the following predicates must be "
                          "met: Number 3 doesn't equal 2")
     
     def test_two_true(self):
         environ = {'test_number': 4}
         p = predicates.Any(EqualsFour(), GreaterThan(3))
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
     
     def test_two_false(self):
         environ = {'test_number': 1}
         p = predicates.Any(EqualsFour(), GreaterThan(3))
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error,
+        self.eval_unmet_predicate(p, environ, 
                          "At least one of the following predicates must be "
                          "met: Number 1 doesn't equal 4, 1 is not greater "
                          "than 3")
@@ -173,169 +163,153 @@ class TestAnyPredicate(unittest.TestCase):
     def test_two_mixed(self):
         environ = {'test_number': 5}
         p = predicates.Any(EqualsFour(), GreaterThan(3))
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
 
 
-class TestIsUserPredicate(unittest.TestCase):
+class TestIsUserPredicate(BasePredicateTester):
     
     def test_user_without_credentials(self):
         environ = {}
         p = predicates.is_user('gustavo')
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error, 'The current user must be "gustavo"')
+        self.eval_unmet_predicate(p, environ,
+                                  'The current user must be "gustavo"')
     
     def test_user_without_userid(self):
         environ = {'repoze.what.credentials': {}}
         p = predicates.is_user('gustavo')
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error, 'The current user must be "gustavo"')
+        self.eval_unmet_predicate(p, environ,
+                                  'The current user must be "gustavo"')
     
     def test_right_user(self):
         environ = make_environ('gustavo')
         p = predicates.is_user('gustavo')
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
     
     def test_wrong_user(self):
         environ = make_environ('andreina')
         p = predicates.is_user('gustavo')
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error, 'The current user must be "gustavo"')
+        self.eval_unmet_predicate(p, environ,
+                                  'The current user must be "gustavo"')
 
 
-class TestInGroupPredicate(unittest.TestCase):
+class TestInGroupPredicate(BasePredicateTester):
     
     def test_user_belongs_to_group(self):
         environ = make_environ('gustavo', ['developers'])
         p = predicates.in_group('developers')
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
     
     def test_user_doesnt_belong_to_group(self):
         environ = make_environ('gustavo', ['developers', 'admins'])
         p = predicates.in_group('designers')
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error,
-                         'The current user must belong to the group "designers"')
+        self.eval_unmet_predicate(p, environ,
+                    'The current user must belong to the group "designers"')
 
 
-class TestInAllGroupsPredicate(unittest.TestCase):
+class TestInAllGroupsPredicate(BasePredicateTester):
     
     def test_user_belongs_to_groups(self):
         environ = make_environ('gustavo', ['developers', 'admins'])
         p = predicates.in_all_groups('developers', 'admins')
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
     
     def test_user_doesnt_belong_to_groups(self):
         environ = make_environ('gustavo', ['users', 'admins'])
         p = predicates.in_all_groups('developers', 'designers')
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error,
-                         'The current user must belong to the group "developers"')
+        self.eval_unmet_predicate(p, environ,
+                    'The current user must belong to the group "developers"')
     
     def test_user_doesnt_belong_to_one_group(self):
         environ = make_environ('gustavo', ['developers'])
         p = predicates.in_all_groups('developers', 'designers')
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error,
-                         'The current user must belong to the group "designers"')
+        self.eval_unmet_predicate(p, environ,
+                    'The current user must belong to the group "designers"')
 
 
-class TestInAnyGroupsPredicate(unittest.TestCase):
+class TestInAnyGroupsPredicate(BasePredicateTester):
     
     def test_user_belongs_to_groups(self):
         environ = make_environ('gustavo', ['developers',' admins'])
         p = predicates.in_any_group('developers', 'admins')
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
     
     def test_user_doesnt_belong_to_groups(self):
         environ = make_environ('gustavo', ['users', 'admins'])
         p = predicates.in_any_group('developers', 'designers')
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error,
+        self.eval_unmet_predicate(p, environ,
                          'The member must belong to at least one of the '
                          'following groups: developers, designers')
     
     def test_user_doesnt_belong_to_one_group(self):
         environ = make_environ('gustavo', ['designers'])
         p = predicates.in_any_group('developers', 'designers')
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
 
 
-class TestNotAnonymousPredicate(unittest.TestCase):
+class TestNotAnonymousPredicate(BasePredicateTester):
     
     def test_authenticated_user(self):
         environ = make_environ('gustavo')
         p = predicates.not_anonymous()
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
     
     def test_anonymous_user(self):
         environ = {}
         p = predicates.not_anonymous()
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error,
+        self.eval_unmet_predicate(p, environ,
                          'The current user must have been authenticated')
 
 
-class TestHasPermissionPredicate(unittest.TestCase):
+class TestHasPermissionPredicate(BasePredicateTester):
     
     def test_user_has_permission(self):
         environ = make_environ('gustavo', permissions=['watch-tv'])
         p = predicates.has_permission('watch-tv')
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
     
     def test_user_doesnt_have_permission(self):
         environ = make_environ('gustavo', permissions=['watch-tv'])
         p = predicates.has_permission('eat')
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error, 'The user must have the "eat" permission')
+        self.eval_unmet_predicate(p, environ,
+                                  'The user must have the "eat" permission')
 
 
-class TestHasAllPermissionsPredicate(unittest.TestCase):
+class TestHasAllPermissionsPredicate(BasePredicateTester):
     
     def test_user_has_all_permissions(self):
         environ = make_environ('gustavo', permissions=['watch-tv', 'party',
                                                        'eat'])
         p = predicates.has_all_permissions('watch-tv', 'eat')
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
     
     def test_user_doesnt_have_permissions(self):
         environ = make_environ('gustavo', permissions=['watch-tv', 'party',
                                                        'eat'])
         p = predicates.has_all_permissions('jump', 'scream')
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error, 'The user must have the "jump" permission')
+        self.eval_unmet_predicate(p, environ,
+                                  'The user must have the "jump" permission')
     
     def test_user_has_one_permission(self):
         environ = make_environ('gustavo', permissions=['watch-tv', 'party',
                                                        'eat'])
         p = predicates.has_all_permissions('party', 'scream')
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error, 'The user must have the "scream" permission')
+        self.eval_unmet_predicate(p, environ,
+                                  'The user must have the "scream" permission')
 
 
-class TestUserHasAnyPermissionsPredicate(unittest.TestCase):
+class TestUserHasAnyPermissionsPredicate(BasePredicateTester):
     
     def test_user_has_all_permissions(self):
         environ = make_environ('gustavo', permissions=['watch-tv', 'party',
                                                        'eat'])
         p = predicates.has_any_permission('watch-tv', 'eat')
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
     
     def test_user_doesnt_have_all_permissions(self):
         environ = make_environ('gustavo', permissions=['watch-tv', 'party',
                                                        'eat'])
         p = predicates.has_any_permission('jump', 'scream')
-        self.assertFalse(p.eval_with_environ(environ))
-        self.assertEqual(p.error,
+        self.eval_unmet_predicate(p, environ,
                          'The user must have at least one of the following '
                          'permissions: jump, scream')
     
@@ -343,8 +317,7 @@ class TestUserHasAnyPermissionsPredicate(unittest.TestCase):
         environ = make_environ('gustavo', permissions=['watch-tv', 'party',
                                                        'eat'])
         p = predicates.has_any_permission('party', 'scream')
-        self.assertTrue(p.eval_with_environ(environ))
-        self.assertEqual(p.error, '')
+        self.eval_met_predicate(p, environ)
 
 
 #{ Test utilities
@@ -406,6 +379,5 @@ class LessThan(predicates.Predicate):
     def _eval_with_environ(self, environ):
         self.number = environ.get('test_number')
         return self.number < self.compared_number
-
 
 #}